+++ aws-sdk-core/lib/aws-sdk-core.rb
@@ -57,6 +57,7 @@ module Aws
     Lambda
     LambdaPreview
     MachineLearning
+    MarketplaceCommerceAnalytics
     OpsWorks
     RDS
     Redshift