+++ aws-sdk-core/lib/aws-sdk-core/marketplacecommerceanalytics.rb
@@ -1,0 +1,4 @@
+Aws.add_service(:MarketplaceCommerceAnalytics, {
+  api: "#{Aws::API_DIR}/marketplacecommerceanalytics/2015-07-01/api-2.json",
+  docs: "#{Aws::API_DIR}/marketplacecommerceanalytics/2015-07-01/docs-2.json",
+})